+++ src/VocabularyCoach.Infrastructure.Sqlite/Extensions/TranslationExtensions.cs
@@ -1,17 +1,47 @@
+using System;
+using System.Linq;
 using VocabularyCoach.Infrastructure.Sqlite.Entities;
 using VocabularyCoach.Models;
+using VocabularyCoach.Services.Data;
 
 namespace VocabularyCoach.Infrastructure.Sqlite.Extensions
 {
 	internal static class TranslationExtensions
 	{
-		public static Translation ToModel(this TranslationEntity translationEntity)
+		public static Translation ToModel(this TranslationEntity translationEntity, int studiedLanguageId, int knownLanguageId)
 		{
 			return new Translation
 			{
-				Text1 = translationEntity.Text1.ToModel(),
-				Text2 = translationEntity.Text2.ToModel(),
+				Text1 = GetTextInLanguage(translationEntity, studiedLanguageId).ToModel(),
+				Text2 = GetTextInLanguage(translationEntity, knownLanguageId).ToModel(),
 			};
+		}
+
+		public static StudiedTranslationData ToStudiedTranslationData(this TranslationEntity translationEntity, int studiedLanguageId, int knownLanguageId)
+		{
+			var textInStudiedLanguage = GetTextInLanguage(translationEntity, studiedLanguageId);
+
+			return new StudiedTranslationData
+			{
+				TextInStudiedLanguage = textInStudiedLanguage.ToModel(),
+				TextInKnownLanguage = GetTextInLanguage(translationEntity, knownLanguageId).ToModel(),
+				CheckResults = textInStudiedLanguage.CheckResults.Select(y => y.ToModel()).ToList(),
+			};
+		}
+
+		private static TextEntity GetTextInLanguage(TranslationEntity translationEntity, int languageId)
+		{
+			if (translationEntity.Text1.LanguageId == languageId)
+			{
+				return translationEntity.Text1;
+			}
+
+			if (translationEntity.Text2.LanguageId == languageId)
+			{
+				return translationEntity.Text2;
+			}
+
+			throw new InvalidOperationException($"Translation does not contain text with language {languageId}");
 		}
 
 		public static TranslationEntity ToEntity(this Translation model)