+++ src/VocabularyCoach.Services/Data/UserSettingsData.cs
@@ -1,0 +1,11 @@
+using VocabularyCoach.Models;
+
+namespace VocabularyCoach.Services.Data
+{
+	public class UserSettingsData
+	{
+		public Language LastStudiedLanguage { get; init; }
+
+		public Language LastKnownLanguage { get; init; }
+	}
+}