+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 2.7.0 - 2023.10.28
+- Added saving of selected languages to user settings.
+- Added thousand separator for values in statistics chart.
+
 ## 2.6.0 - 2023.09.25
 - Added masking of same words for synonyms.